fix: keep mermaid zoom test under file budget
